requirements-dev.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg==3.1.11
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
